+++ libraries/MAX520/CHANGELOG.md
@@ -6,5 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-03-07
+- improve documentation
+- fix reset() value byte
+- fix example addresses
+- add examples
+- update keywords.txt
+- update unit_test
+
+
 ## [0.1.0] - 2024-03-06
 - initial version